[casper_files/r2dbe_top.slx] VDIF headers debugged
This is the first complete working version of the r2dbe. I am
just transmitting a counter as data from the VDIF block, and
the software still does not to ADC calibration.  The software
is set up for only one ten gbe port at the moment.

This is tested and correctly reads and sets the headers.

Next test is to debug by plugging in a 1 PPS signal and noise
generator in lab.


Former-commit-id: e55989bc6dbd78f0e69faf3d15050c5d707ea78b