[casper_files/r2dbe_vdif_lib.slx] changed W4 and W5 of vdif headers to pure 32 bit software registers
W4-W7 are free to assign, for now we are using the ALMA standard
EDV 2.  W6 and W7 combine to make a 64 bit "Packet Serial Number"
aka PSN.  This is fine, and I chose to leave those in the code
as they are counters dependent on an internal start-of-packet
signal.  However, words 4 and 5 should really just be software
registers so these can be reassigned in software without
recompiling.


Former-commit-id: eaaa6e37bc90797ff5d7849325b00c23806c5d9c