[r2dbe_top, r2dbe_vdif_lib] Changed the 1024 to 1028 in w2 of the vdif hdr block
After the design review at SAO on Aug 13, it was brought to our
attention that the 1024 we coded into the design for the vdif
header's data frame length number should also include the length
of the header.  The appropriate number was calculated to be 1028.

After this change and compile (...Aug_13_1405.bof.gz), all soft-
ware checks of the data recorded to the mark 6 show pass.
Previously, dqa showed garbage because it used this 1024 number
to find the location of the next header in the data file, and
was erroneously interpretting data as header values.  Now
the values are as expected.


Former-commit-id: 23e492f8346b24e23dbedb00e7dfae9204eedf1f