[software/VDIF_test.py] arming earlier
Just after the ADC MMCM calibration is done, the one pps shoud
arm to the GPS pps.  Then a sleep for 2 seconds so that the
internal PPS signal stabilizes.

Now that there is a stable internal pps, we can proceed to
initialize the rest of the system including headers.

An issue with this is that occasionally if the GPS pps varies
in period the startup internal pps signal could have a shortened
or extended period when arming to a new GPS pps.  This has the
effect that on some startups the packets streaming would be
stamped with a time up to a second late.  This fixes that issue
and has produced reliable time stamps.


Former-commit-id: 51911d0bbbb2d8e649bfb9a3f152ab79ae516a8b