[software/capture_32.sh] adding script for packet capture
I am adding a script that will allow the root user to capture
from tcpdump data from the R2DBE associated with the start of the
second.

This listens and captures 32 consecutive packets starting at
data frame 0 (at the start of a new second in time) on ports eth3
and eth5 of the Mark6.

This piece of code is useful for zero baseline tests and observing
signal statistics out of the R2DBE.


Former-commit-id: 70e582eee40596e2aba0f41f453820b6b4acd36a