[software/VDIF_test.py] uses netifaces to grab relevant ip addresses
No longer does the user need to manually edit the file and add
ip addresses of eth3 and eth5, the python code has been smartened
up to grab those interfaces using netifaces.

One thing could still be improved and it is the fact that the MAC
address obtained through this method is not guaranteed to be the
full 48 bits expected by the tengbe_config script.

Initial testing shows that this has no ill effect, however.


Former-commit-id: fc730f836f5447d1e8c9ec257636f89b35baaa07