[r2dbe_top, r2dbe_rev2.bof.gz] Bitcode with counter implemented
I've copied the work I've done on r2dbe_top_ctr.slx to here.
The last commit actually included changes to r2dbe_adc_lib, which
changes the onepps block to have actually the offset counter
implemented and sent out to a software register and ported
out of the block (and into the VDIF header blocks to be put in
W5).

The offset block actually computes difference in edges, like a
stopwatch.  It tracks the offset between the incoming GPS pps
and the internally generated pps (clock derived from the maser).
The output of this block is number of clock cycles between edges.

If the R2DBE pps edge is EARLIER than the GPS pps, the sample
offset will be positive.  There is a 22 sample (~80 ns) loop between
when the internal pps edge is compared to the GPS, and when it is
sent out of the smon port.  Thus comparing the values to a scope
of the pps output and gps pps led to an 80 ns offset.

However, this will measure small fluctuations from second to second
well, and will stamp that into the data. It should record an offset
that is on the order of hundreds of nanoseconds or greater.  So
it's really good down to the tens of nanoseconds, but fringe search
will find that.

This version of the model file still has library links, but those
should really be disabled for future bitcodes.


Former-commit-id: dd8a0d393ef91cf874e6a58640ed2862f1e86c93