[quickspecR2DBE.py] adding Jan Wagner's diagnostic file
Jan Wagner wrote a script to show autocorrelations from the 8-bit
snapshots from each ADC on the R2DBE.  This is very useful as it
involves no capture of packets coming off the R2DBE (which
requires capture_32.sh and another python script, requires being
root, etc).  This is a fast way to tell you that you have plugged
in to the inputs as expected, and get a glimpse at the bandpass
shape.


Former-commit-id: 3c0286288d34e8ebbd7c6a73e7f067705db92d60