[r2dbe_snaps] fixed little end and sample swap
previously the 2 bit snapshot decoder function borrowed code that
interpretted the vdif files.  At that point in  the signal chain,
the data had been byte-swapped to little endian, and the 2-bit
data were reordered in 16 sample groups.

However, the 2 bit snapshot grabs the data before any of these
operations, so the code had to be changed to read big-endian
data and adjust the "shift_by" parameter so that the 2 bit data
came out in the correct order.

This bug was noticed when I looked at the autocorrelation of the
8-bit data vs the 2-bit data.  While the band shape was similar
(low edge rolls off at DC, high edge stays high b/c noise source
BW is wider than Nyquist) but the injected tone showed up on the
opposite side of the Fs/4 tone, which clearly showed I had
manipulated the data incorrectly.


Former-commit-id: 2fe145c00f07ba422977a7d77cd84cba705a030e