Upgrade 2-bit quantizer
[r2dbe_adc_lib.slx] The 2-bit quantizer now has separate thresholds
for positive and negative input 8-bit values, as well as a zero
value. Dithering is applied to zero-inputs. The quantizer is based
on the design recently implemented on the SWARM DBE.

The threshold software register takes a 32-bit input. The msb
selects whether to use the new ('1') or old ('0') 2-bit quantizer.
The old 2-bit quantizer works the same as before, with the lower 7
bits defining both positive and negative thresholds. The new 2-bit
quantizer takes the negative threshold from bits 0-7, the zero value
from bits 8-15, and the positive threshold from bits 16-23.

[r2dbe_top.slx] The bitcode includes the new 2-bit quantizer. In
order to meet timing the I/O delay to 8-bit and 2-bit snap blocks
has been doubled (from 3 to 6).


Former-commit-id: 4ccbb04aab62b252f56e68a20653c098c97185d5