Add non-X version of R2DBE monitor
The non-X version is entirely contained in r2dbe_monitor_nox.py and
can be run as a standalone program. Calling r2dbe_monitor.py with the
--no-X option will compile a suitable list of arguments and use
subprocess.check_call to execute r2dbe_monitor_nox.py.

In the future the monitor may be cleaned up to just run as a single
script.


Former-commit-id: 1c8aec361154934afccf833600a0529b73787212