[lib/mark6] Define a number of helper classes
ModuleStatus encapsulates the response of mstat? query.

Response is a base class that encapsulates a response in da-client
queries and commands. QueryResponse and CommandResponse are
derivative classes that encapsulate query and command responses,
respectively.

InputStream encapsulates an input stream definition.

Mark6Config encapsulates the configuration of a Mark6, both the
class instance and underlying device.

Some useful definitions are also added in defines.py.